+++ src/us/kbase/auth2/lib/identity/IdentityProvider.java
@@ -40,6 +40,8 @@ public interface IdentityProvider {
 	/** Get a set of identities from an identity provider given an identity provider authcode.
 	 * @param authcode the authcode returned from the identity provider on the redirect after
 	 * login.
+	 * @param pkceCodeVerifier the OAuth2 PKCE code verifier.
+	 * See https://www.oauth.com/oauth2-servers/pkce/authorization-request/
 	 * @param link whether the authcode was associated with a login or link url.
 	 * @param environment the name of the environment that was used when configuring the redirect
 	 * url. Pass null for the default environment.
@@ -47,7 +49,8 @@ public interface IdentityProvider {
 	 * @throws IdentityRetrievalException if getting the idenities failed.
 	 * @throws NoSuchEnvironmentException if there is no such environment configured. 
 	 */
-	Set<RemoteIdentity> getIdentities(String authcode, boolean link, String environment)
+	Set<RemoteIdentity> getIdentities(
+			String authcode, String pkceCodeVerifier, boolean link, String environment)
 			throws IdentityRetrievalException, NoSuchEnvironmentException;
 	
 	/** Get the names of the additional environments beyond the default environment that are