+++ src/us/kbase/test/auth2/providers/OrcIDIdentityProviderTest.java
@@ -198,9 +198,9 @@ public class OrcIDIdentityProviderTest {
 	@Test
 	public void illegalAuthcode() throws Exception {
 		final IdentityProvider idp = new OrcIDIdentityProvider(CFG);
-		failGetIdentities(idp, null, true, new IllegalArgumentException(
+		failGetIdentities(idp, null, "pkce", true, new IllegalArgumentException(
 				"authcode cannot be null or empty"));
-		failGetIdentities(idp, "  \t  \n  ", true, new IllegalArgumentException(
+		failGetIdentities(idp, "  \t  \n  ", "pkce", true, new IllegalArgumentException(
 				"authcode cannot be null or empty"));
 	}
 	
@@ -208,27 +208,31 @@ public class OrcIDIdentityProviderTest {
 	public void noSuchEnvironment() throws Exception {
 		final IdentityProvider idp = new OrcIDIdentityProvider(CFG);
 		
-		failGetIdentities(idp, "foo", true, "myenv1", new NoSuchEnvironmentException("myenv1"));
-		failGetIdentities(idp, "foo", false, "myenv1", new NoSuchEnvironmentException("myenv1"));
+		failGetIdentities(idp, "foo", "pkce", true, "myenv1",
+				new NoSuchEnvironmentException("myenv1"));
+		failGetIdentities(idp, "foo", "pkce", false, "myenv1",
+				new NoSuchEnvironmentException("myenv1"));
 	}
 	
 	
 	private void failGetIdentities(
 			final IdentityProvider idp,
 			final String authcode,
+			final String pkce,
 			final boolean link,
 			final Exception exception) throws Exception {
-		failGetIdentities(idp, authcode, link, null, exception);
+		failGetIdentities(idp, authcode, pkce, link, null, exception);
 	}
 	
 	private void failGetIdentities(
 			final IdentityProvider idp,
 			final String authcode,
+			final String pkce,
 			final boolean link,
 			final String env,
 			final Exception exception) throws Exception {
 		try {
-			idp.getIdentities(authcode, link, env);
+			idp.getIdentities(authcode, pkce, link, env);
 			fail("got identities with bad setup");
 		} catch (Exception e) {
 			TestCommon.assertExceptionCorrect(e, exception);
@@ -262,15 +266,15 @@ public class OrcIDIdentityProviderTest {
 				new IdentityRetrievalException("No access token was returned by OrcID");
 		
 		setUpCallAuthToken(acode, null, redir, cliid, clisec, "name", "fake ID");
-		failGetIdentities(idp, acode, false, e);
+		failGetIdentities(idp, acode, "pkce", false, e);
 		setUpCallAuthToken(acode, "\t  ", redir, cliid, clisec, "name", "fake ID");
-		failGetIdentities(idp, acode, false, e);
+		failGetIdentities(idp, acode, "pkce", false, e);
 		
 		setUpCallAuthToken(acode, "fake token", redir, cliid, clisec, "my name", null);
-		failGetIdentities(idp, acode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, acode, "pkce", false, new IdentityRetrievalException(
 				"No id was returned by OrcID"));
 		setUpCallAuthToken(acode, "fake token", redir, cliid, clisec, "my name", "   \t  \n  ");
-		failGetIdentities(idp, acode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, acode, "pkce", false, new IdentityRetrievalException(
 				"No id was returned by OrcID"));
 	}
 	
@@ -284,37 +288,37 @@ public class OrcIDIdentityProviderTest {
 		final String authCode = "foo10";
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, APP_JSON, 200, "foo bar");
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: Unable to parse response from OrcID service."));
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, "text/html", 200,
 				"{\"access_token\":\"foobar\"}");
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: Unable to parse response from OrcID service."));
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, APP_JSON, 500, STRING1000);
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: Got unexpected HTTP code and unparseable response " +
 				"from OrcID service: 500. Response: " + STRING1000));
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, APP_JSON, 500, STRING1001);
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: Got unexpected HTTP code and unparseable response " +
 				"from OrcID service: 500. Truncated response: " + STRING1000));
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, APP_JSON, 500, null);
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: Got unexpected HTTP code with no response body " +
 				"from OrcID service: 500."));
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, APP_JSON, 500, "{}");
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: Got unexpected HTTP code with no error in the " +
 				"response body from OrcID service: 500."));
 		
 		setUpCallAuthToken(authCode, redir, cliid, clisec, APP_JSON, 500,
 				"{\"error\":\"whee!\",\"error_description\":\"whoo!\"}");
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Authtoken retrieval failed: OrcID service returned an error. HTTP code: 500. " +
 				"Error: whee!. Error description: whoo!"));
 	}
@@ -332,44 +336,44 @@ public class OrcIDIdentityProviderTest {
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, APP_JSON, 200, "bleah");
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Unable to parse response from OrcID service."));
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, "text/html", 200, MAPPER.writeValueAsString(
 				map("id", "id1", "displayName", "dispname1", "emails", Arrays.asList(
 						map("value", "email1")))));
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Unable to parse response from OrcID service."));
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, APP_JSON, 500, STRING1000);
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Got unexpected HTTP code and unparseable " +
 				"response from OrcID service: 500. Response: " + STRING1000));
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, APP_JSON, 500, STRING1001);
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Got unexpected HTTP code and unparseable " +
 				"response from OrcID service: 500. Truncated response: " + STRING1000));
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, APP_JSON, 500, null);
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Got unexpected HTTP code with no response " +
 				"body from OrcID service: 500."));
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, APP_JSON, 500, "{}");
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"Got unexpected HTTP code with no error in " +
 				"the response body from OrcID service: 500."));
 		
 		setUpCallAuthToken(authCode, authtoken, redir, cliid, clisec, "my name", orcID);
 		setupCallID(authtoken, orcID, APP_JSON, 401, MAPPER.writeValueAsString(
 				map("error", "whee!", "error_description", "whoo!")));
-		failGetIdentities(idp, authCode, false, new IdentityRetrievalException(
+		failGetIdentities(idp, authCode, "pkce", false, new IdentityRetrievalException(
 				"OrcID service returned an error. HTTP code: 401. " +
 				"Error: whee!. Error description: whoo!"));
 	}
@@ -395,7 +399,7 @@ public class OrcIDIdentityProviderTest {
 		setUpCallAuthToken(authCode, "footoken3", "https://ologinredir.com",
 				idconfig.getClientID(), idconfig.getClientSecret(), " My name ", orcID);
 		setupCallID("footoken3", orcID, APP_JSON, 200, MAPPER.writeValueAsString(response));
-		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, false, null);
+		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, "pkce", false, null);
 		assertThat("incorrect number of idents", rids.size(), is(1));
 		final Set<RemoteIdentity> expected = new HashSet<>();
 		expected.add(new RemoteIdentity(new RemoteIdentityID(ORCID, orcID),
@@ -414,7 +418,7 @@ public class OrcIDIdentityProviderTest {
 				idconfig.getClientID(), idconfig.getClientSecret(), " My name ", orcID);
 		setupCallID("footoken3", orcID, APP_JSON, 200, MAPPER.writeValueAsString(
 				map("email", Arrays.asList(map("email", "email7")))));
-		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, false, "e3");
+		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, "pkce", false, "e3");
 		assertThat("incorrect number of idents", rids.size(), is(1));
 		final Set<RemoteIdentity> expected = new HashSet<>();
 		expected.add(new RemoteIdentity(new RemoteIdentityID(ORCID, orcID),
@@ -453,7 +457,7 @@ public class OrcIDIdentityProviderTest {
 				null, orcID);
 		setupCallID("footoken2", orcID, APP_JSON, 200, MAPPER.writeValueAsString(
 				response));
-		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, true, null);
+		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, "pkce", true, null);
 		assertThat("incorrect number of idents", rids.size(), is(1));
 		final Set<RemoteIdentity> expected = new HashSet<>();
 		expected.add(new RemoteIdentity(new RemoteIdentityID(ORCID, orcID),
@@ -482,7 +486,7 @@ public class OrcIDIdentityProviderTest {
 				null, orcID);
 		setupCallID("footoken2", orcID, APP_JSON, 200, MAPPER.writeValueAsString(
 				map("email", Arrays.asList(map("email", "email4")))));
-		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, true, "e3");
+		final Set<RemoteIdentity> rids = idp.getIdentities(authCode, "pkce", true, "e3");
 		assertThat("incorrect number of idents", rids.size(), is(1));
 		final Set<RemoteIdentity> expected = new HashSet<>();
 		expected.add(new RemoteIdentity(new RemoteIdentityID(ORCID, orcID),